Fix more bugs in game-over handling
There really does need to be a check in next_turn() because that's how we
signal a !done at retaliation time. The rule is, if the current player is
in retaliation state, and next_turn is called, they finally transition to dead.

Any subclassed game.player_dead, not self.player_dead is not used as we do
not signal retaliation -> death for printing a message or anything.